+++ apps/desktop-wallet/src/components/AssetLogo.tsx
@@ -18,10 +18,10 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 
 import { FungibleToken, NFT } from '@alephium/shared'
 import { HelpCircle } from 'lucide-react'
+import ReactPlayer from 'react-player'
 import styled from 'styled-components'
 
 interface AssetLogoProps {
-  assetId: FungibleToken['id']
   assetImageUrl: FungibleToken['logoURI'] | NFT['image']
   size: number
   assetName?: FungibleToken['name']
@@ -29,9 +29,11 @@ interface AssetLogoProps {
   className?: string
 }
 
-const AssetLogo = ({ assetId, assetImageUrl, size, assetName, className }: AssetLogoProps) => (
+const AssetLogo = ({ assetImageUrl, size, assetName, className }: AssetLogoProps) => (
   <div className={className}>
-    {assetImageUrl ? (
+    {assetImageUrl?.endsWith('.mp4') ? (
+      <ReactPlayer url={assetImageUrl} muted width={size} height={size} />
+    ) : assetImageUrl ? (
       <LogoImage src={assetImageUrl} />
     ) : assetName ? (
       <Initials size={size}>{assetName.slice(0, 2)}</Initials>